Rename block_tech_shift_reservations to sync_booked_blackouts and add removal logic
- Renamed command from block_tech_shift_reservations to sync_booked_blackouts
- Updated shift times to correct hours: AM shift 10am-4pm, PM shift 4pm-10pm
- Added functionality to remove existing blackouts when techs are now staffed
- Command now performs full sync: creates needed blackouts AND removes unnecessary ones
- Updated all tests to match new functionality
- Fixed timezone handling for blackout comparison
