+++ protohaven_api/cli.py
@@ -5,6 +5,7 @@ import logging
 import sys
 
 from protohaven_api.commands import (
+    blackouts,
     classes,
     clearances,
     comms,
@@ -13,7 +14,6 @@ from protohaven_api.commands import (
     maintenance,
     reservations,
     roles,
-    tech_shifts,
     violations,
 )
 from protohaven_api.config import get_config
@@ -43,6 +43,7 @@ else:
 
 
 class ProtohavenCLI(  # pylint: disable=too-many-ancestors
+    blackouts.Commands,
     comms.Commands,
     reservations.Commands,
     classes.Commands,
@@ -52,7 +53,6 @@ class ProtohavenCLI(  # pylint: disable=too-many-ancestors
     roles.Commands,
     maintenance.Commands,
     clearances.Commands,
-    tech_shifts.Commands,
 ):
     """argparser-based CLI for protohaven operations"""
