--- protohaven_api/commands/tech_shifts.py
@@ -1,196 +1,0 @@
-"""Commands related to tech shift management and blackouts"""
-
-import argparse
-import datetime
-import logging
-from typing import Optional
-
-from protohaven_api.automation.techs import techs as forecast
-from protohaven_api.commands.decorator import arg, command, print_yaml
-from protohaven_api.config import tznow
-from protohaven_api.integrations import booked
-
-log = logging.getLogger("cli.tech_shifts")
-
-
-class Commands:
-    """Commands for managing tech shifts and related blackouts"""
-
-    @command(
-        arg(
-            "--apply",
-            help="Actually create blackouts (dry run if false)",
-            action=argparse.BooleanOptionalAction,
-            default=False,
-        ),
-        arg(
-            "--days-ahead",
-            help="Number of days to check ahead for tech shifts",
-            type=int,
-            default=14,
-        ),
-        arg(
-            "--start-date",
-            help="Start date for checking tech shifts (YYYY-MM-DD format, defaults to today)",
-            type=str,
-            default=None,
-        ),
-        arg(
-            "--block-full-day",
-            help="Block reservations for entire day when no AM nor PM shift (default: true)",
-            action=argparse.BooleanOptionalAction,
-            default=True,
-        ),
-    )
-    def block_tech_shift_reservations(self, args, _):
-        """Check tech shift forecast and block tool reservations for shifts with no techs.
-        
-        For each day in the forecast period:
-        - If a shift (AM or PM) has 0 techs scheduled, block all tool reservations for that shift
-        - If both AM and PM shifts have 0 techs (or --block-full-day is true), block entire day
-        """
-        # Parse start date
-        if args.start_date:
-            try:
-                start_date = datetime.datetime.strptime(args.start_date, "%Y-%m-%d").date()
-                start_date = datetime.datetime.combine(start_date, datetime.time.min)
-            except ValueError:
-                log.error(f"Invalid start date format: {args.start_date}. Use YYYY-MM-DD")
-                return
-        else:
-            start_date = tznow().replace(hour=0, minute=0, second=0, microsecond=0)
-        
-        # Get tech shift forecast
-        log.info(f"Getting tech shift forecast for {args.days_ahead} days starting from {start_date.date()}")
-        forecast_data = forecast.generate(start_date, args.days_ahead, include_pii=False)
-        calendar_view = forecast_data["calendar_view"]
-        
-        # Get all resources/tools from Booked
-        log.info("Fetching all resources from Booked...")
-        all_resources = booked.get_resources()
-        resource_ids = [r["resourceId"] for r in all_resources]
-        log.info(f"Found {len(resource_ids)} resources/tools in Booked")
-        
-        # Get existing blackouts to avoid duplicates
-        existing_blackouts = booked.get_blackouts(
-            start_date=start_date,
-            end_date=start_date + datetime.timedelta(days=args.days_ahead)
-        )
-        existing_blackout_times = set()
-        for blackout in existing_blackouts.get("blackouts", []):
-            start = datetime.datetime.fromisoformat(blackout["startDateTime"].replace("Z", "+00:00"))
-            end = datetime.datetime.fromisoformat(blackout["endDateTime"].replace("Z", "+00:00"))
-            # Convert to naive datetime for comparison (command uses naive datetimes)
-            if start.tzinfo:
-                start = start.replace(tzinfo=None)
-            if end.tzinfo:
-                end = end.replace(tzinfo=None)
-            existing_blackout_times.add((start, end))
-        
-        blackouts_created = []
-        days_checked = 0
-        
-        for day in calendar_view:
-            date_str = day["date"]
-            date_obj = datetime.datetime.strptime(date_str, "%Y-%m-%d")
-            
-            am_techs = len(day["AM"]["people"])
-            pm_techs = len(day["PM"]["people"])
-            
-            log.info(f"{date_str}: AM shift has {am_techs} techs, PM shift has {pm_techs} techs")
-            
-            # Skip holidays (shop is closed anyway)
-            if day["is_holiday"]:
-                log.info(f"  Skipping holiday: {date_str}")
-                continue
-            
-            days_checked += 1
-            
-            # Define shift times (adjust as needed for your schedule)
-            am_start = date_obj.replace(hour=9, minute=0, second=0)   # 9:00 AM
-            am_end = date_obj.replace(hour=13, minute=0, second=0)    # 1:00 PM
-            pm_start = date_obj.replace(hour=13, minute=0, second=0)  # 1:00 PM
-            pm_end = date_obj.replace(hour=17, minute=0, second=0)    # 5:00 PM
-            
-            # Check if we should block entire day
-            block_entire_day = args.block_full_day and (am_techs == 0 and pm_techs == 0)
-            
-            if block_entire_day:
-                # Block entire day (9 AM to 5 PM)
-                start_time = am_start
-                end_time = pm_end
-                
-                if (start_time, end_time) not in existing_blackout_times:
-                    blackouts_created.append({
-                        "date": date_str,
-                        "type": "full_day",
-                        "reason": "No techs scheduled for AM or PM shift",
-                        "start": start_time.isoformat(),
-                        "end": end_time.isoformat(),
-                        "resource_ids": resource_ids
-                    })
-                    log.info(f"  Will block entire day: {start_time.time()} to {end_time.time()}")
-                else:
-                    log.info(f"  Blackout already exists for entire day")
-            
-            else:
-                # Check individual shifts
-                if am_techs == 0:
-                    if (am_start, am_end) not in existing_blackout_times:
-                        blackouts_created.append({
-                            "date": date_str,
-                            "type": "am_shift",
-                            "reason": "No techs scheduled for AM shift",
-                            "start": am_start.isoformat(),
-                            "end": am_end.isoformat(),
-                            "resource_ids": resource_ids
-                        })
-                        log.info(f"  Will block AM shift: {am_start.time()} to {am_end.time()}")
-                    else:
-                        log.info(f"  Blackout already exists for AM shift")
-                
-                if pm_techs == 0:
-                    if (pm_start, pm_end) not in existing_blackout_times:
-                        blackouts_created.append({
-                            "date": date_str,
-                            "type": "pm_shift",
-                            "reason": "No techs scheduled for PM shift",
-                            "start": pm_start.isoformat(),
-                            "end": pm_end.isoformat(),
-                            "resource_ids": resource_ids
-                        })
-                        log.info(f"  Will block PM shift: {pm_start.time()} to {pm_end.time()}")
-                    else:
-                        log.info(f"  Blackout already exists for PM shift")
-        
-        # Apply blackouts if --apply is set
-        if args.apply and blackouts_created:
-            log.info(f"\nCreating {len(blackouts_created)} blackout(s)...")
-            for blackout in blackouts_created:
-                try:
-                    result = booked.create_blackout(
-                        start_date=datetime.datetime.fromisoformat(blackout["start"]),
-                        end_date=datetime.datetime.fromisoformat(blackout["end"]),
-                        resource_ids=blackout["resource_ids"],
-                        title=f"Tech Shift Blackout - {blackout['date']} {blackout['type']}",
-                        description=f"Automated blackout: {blackout['reason']}"
-                    )
-                    log.info(f"  Created blackout: {blackout['date']} {blackout['type']} (ID: {result.get('blackoutId', 'N/A')})")
-                except Exception as e:
-                    log.error(f"  Failed to create blackout for {blackout['date']} {blackout['type']}: {e}")
-        elif blackouts_created:
-            log.info(f"\nDry run: Would create {len(blackouts_created)} blackout(s)")
-            for blackout in blackouts_created:
-                log.info(f"  Would block {blackout['date']} {blackout['type']}: {blackout['start']} to {blackout['end']}")
-        else:
-            log.info(f"\nNo blackouts needed for {days_checked} days checked")
-        
-        # Output summary
-        summary = {
-            "days_checked": days_checked,
-            "blackouts_needed": len(blackouts_created),
-            "blackouts_created": blackouts_created if args.apply else [],
-            "dry_run": not args.apply,
-        }
-        
-        print_yaml(summary)
